requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==2.1.1
db-dtypes==1.1.1
google==3.0.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud==0.34.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
linkedin-scraper==2.9.1
lxml==4.9.2
numpy==1.26.1
packaging==23.2
pandas==2.1.1
proto-plus==1.22.3
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==1.26.17
